requirements.txt
# Core AI/ML dependencies
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Multimodal models
Pillow>=10.0.0
opencv-python>=4.8.0

# Vector database
chromadb>=0.4.15

# LangGraph for agentic workflows
langgraph>=0.0.40
langchain>=0.1.0
langchain-community>=0.0.20

# API and web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# CLIP for image embeddings (using torch-compatible version)
open-clip-torch>=2.0.0

# Additional ML dependencies
torchvision>=0.16.0
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
requests>=2.31.0